+++ src/pic/simulator/parser/commands/Incfsz.java
@@ -32,7 +32,7 @@ public class Incfsz extends Command
 
 	@Override
 	public void execute(PicProcessor proc) {
-		byte f = proc.getMemoryControl().getAt(arg0);
+		short f = proc.getMemoryControl().getAt(arg0);
 		f++;
 		
 		if(arg1==0)
@@ -42,7 +42,7 @@ public class Incfsz extends Command
 
 		if(f==0)
 		{
-			byte pcl = proc.getMemoryControl().getAt(SpecialFunctionRegister.PCL);
+			short pcl = proc.getMemoryControl().getAt(SpecialFunctionRegister.PCL);
 			pcl++;
 			proc.getMemoryControl().setAt(SpecialFunctionRegister.PCL, pcl);
 		}
